refactor(partners): suffix all SDK entrypoints .logic-function.ts
