+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-application-created.logic-function.ts
