+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-application-set-name.logic-function.ts
