+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-opportunity-partner-assigned.logic-function.ts
