+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-opportunity-partner-won.logic-function.ts
