+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-partner-content-created.logic-function.ts
