+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-partner-link-created.logic-function.ts
