+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/on-partner-service-created.logic-function.ts
