+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/post-install.logic-function.ts
